requirements.txt
flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
line-bot-sdk==3.5.1
psycopg2-binary==2.9.11
sqlalchemy==2.0.30
alembic==1.13.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
anthropic==0.26.0
sendgrid==6.11.0
cryptography==42.0.7
apscheduler==3.10.4
Pillow==12.1.1
flask-sqlalchemy==3.1.1
openai==2.30.0
requests==2.31.0
beautifulsoup4==4.14.3
fpdf2==2.8.7